[update] dependencies and fix the literal declaration.

+++ src/Lamscript/Main.cpp
@@ -13,7 +13,7 @@ int main(int argc, char** argv) {
   lamscript::Expression* expression = new lamscript::Binary(
       new lamscript::Unary(
           lamscript::Token{lamscript::MINUS, "-", nullptr, 1},
-          new lamscript::Literal(static_cast<double>(123))),
+          new lamscript::Literal(123)),
       lamscript::Token{lamscript::STAR, "*", nullptr, 1},
       new lamscript::Grouping(
           new lamscript::Literal(45.67)));
